Tests around Indexer info's interaction with the null prone ClusterPeer. #79

+++ test/java/com/splunk/ClusterPeersTest.java
@@ -42,7 +42,7 @@ public class ClusterPeersTest {
 		ClusterPeer clusterPeer = clusterPeers.get(slaveGuid);
 
 		assertNotNull(clusterPeer);
-		assertEquals(clusterPeer.getPort(), slavePort);
+		assertEquals((int) clusterPeer.getPort(), slavePort);
 		assertEquals(clusterPeer.getHost(), localHostIp);
 	}
 }
